+++ templates/indices/transactions.go
@@ -74,7 +74,7 @@ var Transactions = Object{
 					"type": "keyword",
 				},
 				"nonce": Object{
-					"type": "double",
+					"type": "long",
 				},
 				"operation": Object{
 					"type": "keyword",
